refactor(#741): refactor entranceDuplicateController
