+++ api/controllers/v1/entrance-duplicate/create-from-duplicate.js
@@ -1,0 +1,47 @@
+const EntranceService = require('../../../services/EntranceService');
+const ErrorService = require('../../../services/ErrorService');
+const RightService = require('../../../services/RightService');
+
+module.exports = async (req, res) => {
+  const hasRight = await sails.helpers.checkRight
+    .with({
+      groups: req.token.groups,
+      rightEntity: RightService.RightEntities.ENTRANCE_DUPLICATE,
+      rightAction: RightService.RightActions.CREATE,
+    })
+    .intercept('rightNotFound', () =>
+      res.serverError(
+        'A server error occured when checking your right to create an entrance duplicate.'
+      )
+    );
+  if (!hasRight) {
+    return res.forbidden(
+      'You are not authorized to create an entrance duplicate.'
+    );
+  }
+  if (
+    !(await sails.helpers.checkIfExists.with({
+      attributeName: 'id',
+      attributeValue: req.param('id'),
+      sailsModel: TEntranceDuplicate,
+    }))
+  ) {
+    return res.badRequest(
+      `Could not find duplicate with id ${req.param('id')}.`
+    );
+  }
+
+  const id = req.param('id');
+
+  const duplicate = await TEntranceDuplicate.findOne(id);
+
+  const { entrance, nameDescLoc } = duplicate.content;
+  try {
+    await EntranceService.createEntrance(entrance, nameDescLoc);
+    await TEntranceDuplicate.destroyOne(id);
+    return res.ok();
+  } catch (e) {
+    ErrorService.getDefaultErrorHandler(res)(e);
+    return false;
+  }
+};
