+++ api/controllers/v1/entrance-duplicate/delete-many.js
@@ -1,0 +1,35 @@
+const ErrorService = require('../../../services/ErrorService');
+const RightService = require('../../../services/RightService');
+
+module.exports = async (req, res) => {
+  const hasRight = await sails.helpers.checkRight
+    .with({
+      groups: req.token.groups,
+      rightEntity: RightService.RightEntities.ENTRANCE_DUPLICATE,
+      rightAction: RightService.RightActions.DELETE_ANY,
+    })
+    .intercept('rightNotFound', () =>
+      res.serverError(
+        'A server error occured when checking your right to delete an entrance duplicate.'
+      )
+    );
+  if (!hasRight) {
+    return res.forbidden(
+      'You are not authorized to delete an entrance duplicate.'
+    );
+  }
+  const idArray = req.param('id');
+  if (!idArray) {
+    return res.badRequest('You must provide the id of the duplicates.');
+  }
+
+  try {
+    await TEntranceDuplicate.destroy({
+      id: idArray,
+    });
+    return res.ok();
+  } catch (err) {
+    ErrorService.getDefaultErrorHandler(res)(err);
+    return false;
+  }
+};
