+++ api/controllers/v1/entrance-duplicate/find-all.js
@@ -1,0 +1,41 @@
+const ControllerService = require('../../../services/ControllerService');
+const ErrorService = require('../../../services/ErrorService');
+const MappingV1Service = require('../../../services/MappingV1Service');
+
+module.exports = async (req, res) => {
+  const sort = `${req.param('sortBy', 'dateInscription')} ${req.param(
+    'orderBy',
+    'ASC'
+  )}`;
+  const limit = req.param('limit', 50);
+  const skip = req.param('skip', 0);
+  try {
+    const duplicates = await TEntranceDuplicate.find()
+      .skip(skip)
+      .limit(limit)
+      .sort(sort)
+      .populate('author');
+    const totalNb = await TEntranceDuplicate.count();
+
+    const params = {
+      controllerMethod: 'EntranceDuplicateController.findAll',
+      limit,
+      searchedItem: 'Entrance duplicates',
+      skip,
+      total: totalNb,
+      url: req.originalUrl,
+    };
+
+    return ControllerService.treatAndConvert(
+      req,
+      null,
+      duplicates,
+      params,
+      res,
+      MappingV1Service.convertToEntranceDuplicateList
+    );
+  } catch (err) {
+    ErrorService.getDefaultErrorHandler(res)(err);
+    return false;
+  }
+};
